Chặn FMA và làm tròn kiểu Python trước khi port pha hình học
Pha hình học ra quyết định bằng so sánh số thực, nên bản Go phải khớp từng bit
chứ không chỉ khớp tới hai chữ số thập phân.

num.Round đi qua chuỗi thập phân thay vì math.Round, vì math.Round làm tròn nửa
ra xa số không trên giá trị đã nhân lên còn Python làm tròn giá trị nhị phân
thật về số chẵn. Khớp Python trên 120.000 phép làm tròn so từng bit.

Đặc tả Go cho phép gộp a*b + c thành FMA, kể cả qua nhiều câu lệnh. Đo trên
máy này: 23% số phép a*1.42 + c bị gộp và lệch Python một đơn vị ở bit cuối.
internal/lint phân tích kiểu toàn module và báo mọi phép nhân số thực chưa bọc
float64(). Nó bắt ngay hai chỗ *1.42 trong SizeItem, vốn qua được mọi vector
chỉ vì làm tròn lên số chẵn nuốt mất sai lệch, trừ khi rơi đúng biên.

+++ layout/size.go
@@ -55,7 +55,7 @@ func SizeItem(tm *text.Measure, cfg Config, kind string, lines []string) (wrappe
 		if kind == "end" {
 			pad = 12
 		}
-		return wl, float64(num.Rnd(max(80, tw*1.42+24+pad))), float64(num.Rnd(max(44, th*1.42+16+pad)))
+		return wl, float64(num.Rnd(max(80, float64(tw*1.42)+24+pad))), float64(num.Rnd(max(44, float64(th*1.42)+16+pad)))
 
 	case "db":
 		wl := tm.Wrap(lines, WrapBudget(cfg, kind))
@@ -82,7 +82,7 @@ func SizeLabel(tm *text.Measure, cfg Config, lines []string) (wrapped []string, 
 	if len(wl) > 0 {
 		lw, lh = tm.Box(wl)
 	}
-	return wl, lw + 2*float64(cfg.LabelPad), lh + 2
+	return wl, lw + float64(2*cfg.LabelPad), lh + 2
 }
 
 // HasText cho biết nội dung một dòng bảng có chữ thật hay không, theo đúng cách
